+++ features/fill-storage/src/main/java/com/ryccoatika/sqatools/fillstorage/ui/manage/ManageViewState.kt
@@ -1,15 +1,18 @@
 package com.ryccoatika.sqatools.fillstorage.ui.manage
 
 import androidx.compose.runtime.Immutable
+import com.ryccoatika.sqatools.fillstorage.core.model.FillStorage
 import com.ryccoatika.sqatools.fillstorage.core.model.Storage
 
 @Immutable
 internal data class ManageViewState(
   val storage: Storage,
+  val fillStorageProgress: FillStorage.Progress?,
 ) {
   companion object {
     val Empty = ManageViewState(
       storage = Storage.Empty,
+      fillStorageProgress = null,
     )
   }
 }